Move all store implementations into minimalkv/stores without breaking current imports.
