+++ minimalkv/stores/_memory/__init__.py
@@ -1,0 +1,4 @@
+from ._dict_store import DictStore
+from ._redis_store import RedisStore
+
+__all__ = ["DictStore", "RedisStore"]
