+++ minimalkv/stores/_net/_azure_block_blob_store/_azurestore_common.py
