+++ minimalkv/stores/_net_common.py
@@ -1,3 +1,4 @@
+# TODO: Rename this module.
 LAZY_PROPERTY_ATTR_PREFIX = "_lazy_"
